+++ simplethreadapi/inref/threadedapinotallowed2_MT.c
@@ -144,7 +144,7 @@ void *childthread(void *threadparm)
 	status = ydb_ci_t(YDB_NOTTP, NULL, "CALL-IN"); CHECK_STATUS("ydb_ci_t()", status, fp);
 	status = ydb_cip_t(YDB_NOTTP, NULL, &callin); CHECK_STATUS("ydb_cip_t()", status, fp);
 	status = ydb_ci_tab_open_t(YDB_NOTTP, NULL, "", NULL); CHECK_STATUS("ydb_ci_tab_open_t()", status, fp);
-	status = ydb_ci_tab_switch_t(YDB_NOTTP, NULL, (uint64_t)NULL, NULL); CHECK_STATUS("ydb_ci_tab_switch_t()", status, fp);
+	status = ydb_ci_tab_switch_t(YDB_NOTTP, NULL, (uintptr_t)NULL, NULL); CHECK_STATUS("ydb_ci_tab_switch_t()", status, fp);
 	status = ydb_file_id_free_t(YDB_NOTTP, NULL, NULL); CHECK_STATUS("ydb_file_id_free_t()", status, fp);
 	status = ydb_file_is_identical_t(YDB_NOTTP, NULL, &fileid1, &fileid2); CHECK_STATUS("ydb_file_is_identical_t()", status, fp);
 	status = ydb_file_name_to_id_t(YDB_NOTTP, NULL, &filename, &fileid1); CHECK_STATUS("ydb_file_name_to_id_t()", status, fp);
